correct hardware defs for 328p uart (from salwow uart refactoring); add explicit casts to satisfy newer, finickier compiler; create frozone project for Eliot's Implausibles Halloween costume
git-svn-id: https://svn.circlemud.org/svn/rocket/trunk@739 46170a15-b891-4bf7-af97-fd374ce181e0

+++ src/lib/numeric_input.c
@@ -97,7 +97,7 @@ void ni_start_input(NumericInputAct *act)
 	act->cur_value.mantissa = 0;
 	act->cur_value.neg_exponent = 0;
 
-	RectRegion rr = { &act->region.bbuf, 1, act->region.x+act->region.xlen-1, 1 };
+	RectRegion rr = { &act->region.bbuf, 1, (uint8_t) (act->region.x+act->region.xlen-1), 1 };
 	cursor_show(&act->cursor, rr);
 	act->decimal_present = FALSE;
